+++ pytsite/content/_console_commands.py
@@ -1,6 +1,6 @@
 """Content Console Commands.
 """
-from random import shuffle as _shuffle, random as _random
+from random import shuffle as _shuffle, random as _random, randint as _randint
 import requests as _requests
 from pytsite import image as _image, auth as _auth, console as _console, lang as _lang, events as _events
 from . import _functions
@@ -31,7 +31,7 @@ class Generate(_console.command.Abstract):
         """Print usage info.
         """
         _console.print_info('Usage: content:generate [--num=NUM] [--title-len=LEN] [--lang=LANG] [--no-html] [--short] '
-                            '--model=MODEL --author=LOGIN')
+                            '[--author=LOGIN] --model=MODEL ')
 
     def execute(self, **kwargs: dict):
         """Execute teh command.
@@ -41,17 +41,17 @@ class Generate(_console.command.Abstract):
             self.usage()
             return -1
 
+        # Author
+        author = None
         author_login = kwargs.get('author')
-        if not author_login:
-            self.usage()
-            return -1
+        if author_login:
+            author = _auth.get_user(author_login)
+            if not author:
+                raise _console.Error("'{}' is not a registered user.".format(author_login))
 
+        # Checking if the content model registered
         if not _functions.is_model_registered(model):
             raise _console.Error("'{}' is not a registered content model.".format(model))
-
-        author = _auth.get_user(author_login)
-        if not author:
-            raise _console.Error("'{}' is not a registered user.".format(author_login))
 
         num = int(kwargs.get('num', 10))
         language = kwargs.get('lang', _lang.get_current_lang())
@@ -70,11 +70,35 @@ class Generate(_console.command.Abstract):
                 sections.append(section)
                 _console.print_info(_lang.t('pytsite.content@new_section_created', {'title': title}))
 
-        for m in range(0, num):
-            title = self._generate_title(int(kwargs.get('title-len', 7)))
-            description = self._generate_title()
-            _shuffle(sections)
+        # Generate tags
+        tags = list(_functions.get_tags(language=language))
+        if len(tags) < 10:
+            for n in range(0, 10):
+                tag = _functions.create_tag(self._generate_title(1), language=language)
+                tags.append(tag)
 
+        # Generate content entities
+        for m in range(0, num):
+            # Author
+            if not author_login:
+                user_finder = _auth.find_users()
+                users_count = user_finder.count()
+                if not users_count:
+                    raise _lang.t('pytsite.content@no_users_found')
+
+                author = user_finder.skip(_randint(0, users_count - 1)).first()
+
+            # Title
+            title = self._generate_title(int(kwargs.get('title-len', 7)))
+
+            # Description
+            description = self._generate_title()
+
+            # Preparing sections and tags
+            _shuffle(sections)
+            _shuffle(tags)
+
+            # Body and images
             body = []
             images = [_image.create(self.lp_url)]
             for n in range(2, 3 if short else 6):
@@ -86,21 +110,12 @@ class Generate(_console.command.Abstract):
                     body.append('\n<p>[img:{}]</p>\n'.format(n))
                     body.append(_requests.get(self.li_url).content.decode('utf-8'))
 
-            tags = list(_functions.get_tags(language=language))
-            if len(tags) < 10:
-                for n in range(0, 10):
-                    tag = _functions.create_tag(self._generate_title(1), language=language)
-                    tags.append(tag)
-
-            _shuffle(tags)
-            tags = tags[:5]
-
             entity = _functions.create(model)
             entity.f_set('title', title)
             entity.f_set('description', description)
             entity.f_set('body', ''.join(body))
             entity.f_set('images', images)
-            entity.f_set('tags', tags)
+            entity.f_set('tags', tags[:5])
             entity.f_set('author', author)
             entity.f_set('status', 'published')
             entity.f_set('views_count', int(_random() * 1000))