+++ pytsite/odm_ui/_event_handlers.py
@@ -1,11 +1,12 @@
 """ODM UI Manager.
 """
+from pytsite import lang as _lang, odm as _odm, auth as _auth
+from ._model import UIMixin
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-from pytsite import lang as _lang, odm as _odm, auth as _auth
-from ._model import UIMixin
 
 __models = {}