+++ pytsite/odm_ui/_model.py
@@ -1,12 +1,11 @@
 """ODM UI Model.
 """
+from abc import ABC as _ABC, abstractmethod as _abstractmethod
+from pytsite import odm as _odm, router as _router
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
-from abc import ABC as _ABC, abstractmethod as _abstractmethod
-
-from pytsite import odm as _odm
 
 
 class UIMixin(_ABC):
@@ -27,6 +26,7 @@ class UIMixin(_ABC):
         """
         pass
 
+    @staticmethod
     def browser_search(self, finder: _odm.Finder, query: str):
         """Adjust ODM browser finder in search operation.
         """
@@ -52,3 +52,28 @@ class UIMixin(_ABC):
         """Get delete form description.
         """
         pass
+
+
+class Model(_odm.Model, UIMixin):
+    @property
+    def can_be_modified(self) -> bool:
+        from . import _functions
+        return _functions.check_permissions('modify', self.model, self.id)
+
+    @property
+    def can_be_deleted(self) -> bool:
+        from . import _functions
+        return _functions.check_permissions('delete', self.model, self.id)
+
+    def get_delete_url(self, redirect_url: str=None, json=False) -> str:
+        args = {
+            'model': self.model,
+            'ids': str(self.id),
+        }
+
+        if json:
+            args['json'] = 'true'
+        else:
+            args['__form_redirect'] = redirect_url if redirect_url else _router.current_url()
+
+        return _router.ep_url('pytsite.odm_ui.ep.post_d_form', args)